add `zntrack.nodes` and `zntrack.interfaces` entry points, refactor abc into `interfaces.py`
