+++ .gitignore
@@ -161,3 +161,4 @@ cython_debug/
 .vscode/
 
 tmp/
+.DS_Store
